feature[dashboard]: adiciona dashboard repository para melhor organização
Refs: #47
